# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Insole API
  version: 1.3.0
  description: |
    스마트 인솔 MVP의 인증, 기기, 측정 세션, 센서 수신, 실시간 snapshot, 분석 결과 계약.

    1.1.0 변경 요약:
    - Frame Batch schemaVersion 1.1 (프레임별 receivedAt, protocolVersion, dataMode, IMU, flags, batchId)
    - 수신기용 세션 조회 API와 업로드 상태 보고 API (X-Receiver-Key)
    - ADC 상한 4095 (devices.adcMax, 세션 생성 시 스냅샷)
    - sampleRateHz 50/100, sourceType 요청 필드 (기본 DEVICE)
    - SensorPoint.label, 409 SESSION_NOT_MEASURING disposition
    - 분석 결과 rule-v1.2.0 관찰 단계(ObservationLevel)와 센서 share

    1.2.0 변경 요약(추가 전용, DEC-035):
    - 분석 결과 rule-v1.3.0 좌우 신호 비율(loadSharePct)과 스트라이드 시간(strideTimeMs)
    - MeasurementHistoryItem의 최신 결과 요약 지표(모두 nullable)

    1.3.0 변경 요약(추가 전용, DEC-036):
    - 분석 결과 rule-v1.4.0 IMU 기반 정강이 움직임 요약 AnalysisResultResponse.movementSummary(MovementSummary, nullable)
    - 프레임 배치·수신기 API는 변경 없음(수신기의 1.1.0 핀 유효)
servers:
- url: http://localhost:8080
  description: Local backend
...
paths:
  /api/v1/auth/signup:
    post:
      tags:
      - Auth
      summary: 회원가입
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: 가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/auth/signin:
    post:
      tags:
      - Auth
      summary: 로그인
      operationId: signin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigninRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/devices:
    get:
      tags:
      - Devices
      summary: 내 기기 목록
      operationId: listMyDevices
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 기기 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - Devices
      summary: 기기 등록
      operationId: registerDevice
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '201':
          description: 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/sensor-layouts/{version}:
    get:
      tags:
      - SensorLayouts
      summary: 센서 배치 조회
      operationId: getSensorLayout
      security:
      - bearerAuth: []
      parameters:
      - name: version
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: 센서 배치
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorLayoutResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/measurement-sessions:
    post:
      tags:
      - Measurements
      summary: 측정 세션 생성
      operationId: createMeasurementSession
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMeasurementSessionRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    get:
      tags:
      - Measurements
      summary: 내 측정 기록
      operationId: listMeasurementSessions
      security:
      - bearerAuth: []
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/MeasurementStatus'
      - name: from
        in: query
        description: Include sessions created at or after this UTC instant.
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        description: Include sessions created at or before this UTC instant.
        schema:
          type: string
          format: date-time
      - name: minQualityScore
        in: query
        description: Include completed sessions whose data quality score is at least this value.
        schema:
          type: integer
          minimum: 0
          maximum: 100
      - name: patternCode
        in: query
        description: Include sessions containing this observed pattern code.
        schema:
          type: string
          maxLength: 100
      responses:
        '200':
          description: 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementSessionPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/measurement-sessions/{sessionId}:
    get:
      tags:
      - Measurements
      summary: 측정 상세
      operationId: getMeasurementSession
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/measurement-sessions/{sessionId}/start:
    post:
      tags:
      - Measurements
      summary: 측정 시작
      operationId: startMeasurementSession
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 처리 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/measurement-sessions/{sessionId}/complete:
    post:
      tags:
      - Measurements
      summary: 측정 종료 및 분석 예약
      operationId: completeMeasurementSession
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 처리 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/measurement-sessions/{sessionId}/cancel:
    post:
      tags:
      - Measurements
      summary: 측정 취소
      operationId: cancelMeasurementSession
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 처리 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /internal/v1/measurement-sessions:
    get:
      tags:
      - Receiver
      summary: 수신기용 측정 중 세션 목록
      description: status는 MEASURING만 허용합니다. deviceSerial을 주면 좌우 기기 중 하나의 serialNumber가 일치하는 세션만 반환합니다. 빈 items 배열을 허용합니다.
      operationId: listReceiverSessions
      security:
      - receiverApiKey: []
      parameters:
      - name: status
        in: query
        required: true
        schema:
          type: string
          enum:
          - MEASURING
      - name: deviceSerial
        in: query
        required: false
        schema:
          type: string
          minLength: 1
          maxLength: 100
      responses:
        '200':
          description: 세션 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiverSessionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /internal/v1/measurement-sessions/{sessionId}:
    get:
      tags:
      - Receiver
      summary: 수신기용 세션 조회
      description: 수신기가 세션에 배정된 기기 식별자·센서 수·adcMax·sampleRateHz를 확정하고 상태 전이를 폴링하는 데 사용합니다.
      operationId: getReceiverSession
      security:
      - receiverApiKey: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 세션 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiverSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /internal/v1/measurement-sessions/{sessionId}/receiver-status:
    post:
      tags:
      - Receiver
      summary: 수신기 업로드 상태 보고
      description: MEASURING 세션에서만 허용됩니다. best-effort 보고이며 수신기는 실패해도 재시도하지 않습니다.
      operationId: reportReceiverStatus
      security:
      - receiverApiKey: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiverStatusRequest'
      responses:
        '204':
          description: 반영 성공
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /internal/v1/devices/{deviceId}/heartbeat:
    post:
      tags:
      - Receiver
      summary: 기기 상태 갱신
      operationId: heartbeatDevice
      security:
      - receiverApiKey: []
      parameters:
      - name: deviceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceHeartbeatRequest'
      responses:
        '204':
          description: 반영 성공
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    receiverApiKey:
      type: apiKey
      in: header
      name: X-Receiver-Key
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: 요청 형식 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: 접근 권한 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: 자원 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: 현재 상태와 요청이 충돌함
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnprocessableEntity:
      description: 의미 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    FootSide:
      type: string
      enum:
      - LEFT
      - RIGHT
    MeasurementStatus:
      type: string
      enum:
      - CREATED
      - MEASURING
      - PROCESSING
      - COMPLETED
      - CANCELLED
      - FAILED
    SourceType:
      type: string
      description: 자동 판별하지 않습니다. 기본값은 DEVICE이며 시뮬레이터(mock receiver, e2e 스크립트, 개발용 UI)만 SIMULATED를 명시합니다.
      enum:
      - DEVICE
      - SIMULATED
...
    ReceiverUploadState:
      type: string
      enum:
      - STREAMING
      - UPLOADING
      - UPLOAD_COMPLETE
...
    DeviceStatus:
      type: string
      enum:
      - ACTIVE
      - INACTIVE
      - DISCONNECTED
      - CALIBRATION_REQUIRED
...
    ApiError:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - details
      - traceId
      - timestamp
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
    SignupRequest:
      type: object
      additionalProperties: false
      required:
      - email
      - password
      - name
      properties:
        email:
          type: string
          format: email
          pattern: '.*\S.*'
          maxLength: 255
        password:
          type: string
          minLength: 8
          pattern: '.*\S.*'
          maxLength: 100
          description: Must be at most 72 bytes when encoded as UTF-8.
          x-maxUtf8Bytes: 72
        name:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          maxLength: 50
    SigninRequest:
      type: object
      additionalProperties: false
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
          pattern: '.*\S.*'
          maxLength: 255
        password:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          maxLength: 100
          description: Must be at most 72 bytes when encoded as UTF-8.
          x-maxUtf8Bytes: 72
    UserSummary:
      type: object
      additionalProperties: false
      required:
      - userId
      - email
      - name
      - createdAt
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    AuthTokenResponse:
      type: object
      additionalProperties: false
      required:
      - tokenType
      - accessToken
      - expiresInSeconds
      - user
      properties:
        tokenType:
          type: string
          enum:
          - Bearer
        accessToken:
          type: string
        expiresInSeconds:
          type: integer
          minimum: 1
        user:
          $ref: '#/components/schemas/UserSummary'
    RegisterDeviceRequest:
      type: object
      additionalProperties: false
      required:
      - serialNumber
      - displayName
      - footSide
      - sensorCount
      - sensorLayoutVersion
      - firmwareVersion
      properties:
        serialNumber:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          maxLength: 100
        displayName:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          maxLength: 100
        footSide:
          $ref: '#/components/schemas/FootSide'
        sensorCount:
          type: integer
          enum:
          - 6
          - 8
        sensorLayoutVersion:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          maxLength: 50
        firmwareVersion:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          maxLength: 50
        adcMax:
          type: integer
          description: 기기의 ADC 최댓값. 현재 허용값은 4095뿐이며 생략하면 4095입니다.
          enum:
          - 4095
          default: 4095
    DeviceResponse:
      type: object
      additionalProperties: false
      required:
      - deviceId
      - serialNumber
      - displayName
      - footSide
      - sensorCount
      - sensorLayoutVersion
      - firmwareVersion
      - adcMax
      - activeCalibrationVersion
      - status
      - registeredAt
      properties:
        deviceId:
          type: string
          format: uuid
        serialNumber:
          type: string
        displayName:
          type: string
        footSide:
          $ref: '#/components/schemas/FootSide'
        sensorCount:
          type: integer
          enum:
          - 6
          - 8
        sensorLayoutVersion:
          type: string
        firmwareVersion:
          type: string
        adcMax:
          type: integer
          minimum: 1
          description: 기기 등록 시 확정한 ADC 최댓값. V5 이전에 등록된 레거시 기기는 65535로 백필됩니다.
        activeCalibrationVersion:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/DeviceStatus'
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
        lastBatteryPercent:
          type: number
          minimum: 0
          maximum: 100
          nullable: true
          description: 마지막 heartbeat의 배터리 퍼센트. 미보정(255)은 수신기가 null로 보냅니다.
        lastBatteryMv:
          type: integer
          minimum: 0
          nullable: true
        registeredAt:
          type: string
          format: date-time
...
    SensorLayoutResponse:
      type: object
      additionalProperties: false
      oneOf:
      - properties:
          sensorCount:
            enum:
            - 6
          points:
            minItems: 6
            maxItems: 6
      - properties:
          sensorCount:
            enum:
            - 8
          points:
            minItems: 8
            maxItems: 8
      required:
      - version
      - sensorCount
      - points
      properties:
        version:
          type: string
        sensorCount:
          type: integer
          enum:
          - 6
          - 8
        points:
          type: array
          description: Ordered sensor points. Length must equal sensorCount.
          oneOf:
          - minItems: 6
            maxItems: 6
          - minItems: 8
            maxItems: 8
          items:
            $ref: '#/components/schemas/SensorPoint'
    CreateMeasurementSessionRequest:
      type: object
      additionalProperties: false
      required:
      - leftDeviceId
      - rightDeviceId
      - sampleRateHz
      properties:
        leftDeviceId:
          type: string
          format: uuid
        rightDeviceId:
          type: string
          format: uuid
        sampleRateHz:
          type: integer
          description: 전송률의 단일 출처. BLE 전송 50Hz(측정 100Hz 분주) 또는 100Hz. UI 기본값은 50.
          enum:
          - 50
          - 100
        sourceType:
          allOf:
          - $ref: '#/components/schemas/SourceType'
          default: DEVICE
          description: 생략하면 DEVICE. 시뮬레이터만 SIMULATED를 명시합니다.
        memo:
          type: string
          maxLength: 500
          nullable: true
    MeasurementSessionResponse:
      type: object
      additionalProperties: false
      required:
      - sessionId
      - status
      - leftDeviceId
      - rightDeviceId
      - sampleRateHz
      - sourceType
      - adcMax
      - createdAt
      properties:
        sessionId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/MeasurementStatus'
        leftDeviceId:
          type: string
          format: uuid
        rightDeviceId:
          type: string
          format: uuid
        sampleRateHz:
          type: integer
        sourceType:
          $ref: '#/components/schemas/SourceType'
        adcMax:
          type: integer
          minimum: 1
          description: 세션 생성 시 기기에서 스냅샷한 ADC 최댓값.
        memo:
          type: string
          nullable: true
        dataQualityScore:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        receiverState:
          allOf:
          - $ref: '#/components/schemas/ReceiverUploadState'
          nullable: true
          description: 수신기가 마지막으로 보고한 업로드 상태. 보고 전이면 null.
        receiverPendingBatches:
          type: integer
          minimum: 0
          nullable: true
          description: 수신기가 마지막으로 보고한 미전송 배치 수.
    MeasurementSessionPage:
      type: object
      additionalProperties: false
      required:
      - items
      - page
      - size
      - totalElements
      - totalPages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MeasurementHistoryItem'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    DeviceHeartbeatRequest:
      type: object
      additionalProperties: false
      required:
      - receiverId
      - observedAt
      - connected
      properties:
        receiverId:
          type: string
          minLength: 1
          pattern: '.*\S.*'
        observedAt:
          type: string
          format: date-time
        connected:
          type: boolean
        batteryPercent:
          type: number
          minimum: 0
          maximum: 100
          nullable: true
          description: Status battery_pct. 255(미보정/미측정)는 수신기가 null로 보냅니다.
        batteryMv:
          type: integer
          minimum: 0
          nullable: true
          description: Status battery_mv.
        firmwareVersion:
          type: string
          minLength: 1
          maxLength: 50
          nullable: true
          description: Status fw_major.fw_minor.fw_patch (예 0.2.0). 기기 firmwareVersion을 갱신합니다.
        rssi:
          type: integer
          nullable: true
...
    ReceiverSessionResponse:
      type: object
      additionalProperties: false
      required:
      - sessionId
      - status
      - sampleRateHz
      - sourceType
      - adcMax
      - startedAt
      - endedAt
      - left
      - right
      - receiverState
      - receiverPendingBatches
      properties:
        sessionId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/MeasurementStatus'
        sampleRateHz:
          type: integer
          enum:
          - 50
          - 100
        sourceType:
          $ref: '#/components/schemas/SourceType'
        adcMax:
          type: integer
          minimum: 1
          description: 세션 adcMax 스냅샷. left/right.adcMax와 같습니다.
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
        left:
          $ref: '#/components/schemas/ReceiverSessionDevice'
        right:
          $ref: '#/components/schemas/ReceiverSessionDevice'
        receiverState:
          allOf:
          - $ref: '#/components/schemas/ReceiverUploadState'
          nullable: true
        receiverPendingBatches:
          type: integer
          minimum: 0
          nullable: true
    ReceiverSessionListResponse:
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReceiverSessionResponse'
    ReceiverStatusRequest:
      type: object
      additionalProperties: false
      required:
      - receiverId
      - state
      - pendingBatchCount
      - observedAt
      properties:
        receiverId:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          maxLength: 100
        state:
          $ref: '#/components/schemas/ReceiverUploadState'
        pendingBatchCount:
          type: integer
          minimum: 0
        observedAt:
          type: string
          format: date-time
...